revozen/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
langchain
langchain-google-genai
